pyproject.toml
[project]
name = "scythe-benchmark"
version = "0.1.0"
description = "Benchmark suite for Scythe scatter/gather subdivision strategies"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "rich>=15.0.0",
    "scythe-engine>=1.3.0",
]
